feat(view): i18nをRNでもりようできるようにする
