+++ src/expo/_layout.tsx
@@ -3,15 +3,46 @@ import {
     DefaultTheme,
     ThemeProvider,
 } from "@react-navigation/native";
+import { getLocales } from "expo-localization";
 import { Slot } from "expo-router";
+import i18n from "i18next";
+import { useEffect, useState } from "react";
+import { initReactI18next } from "react-i18next";
 import { useColorScheme } from "react-native";
 import { SafeAreaView } from "react-native-safe-area-context";
+import { i18nAppConfig } from "src/locales/i18n";
 import { TamaguiProvider } from "tamagui";
-
 import tamaguiConfig from "tamagui.config";
 
 export default function RootLayout() {
     const colorScheme = useColorScheme();
+    const [languageLoaded, setLanguageLoaded] = useState(false);
+    const [language, setLanguage] = useState<string | null>();
+
+    useEffect(() => {
+        const getSystemLanguageAndSet = async () => {
+            return getLocales()?.[0]?.languageTag ?? "en";
+        };
+
+        getSystemLanguageAndSet().then((phoneLocale) => {
+            setLanguage(phoneLocale);
+        });
+    }, []);
+
+    useEffect(() => {
+        if (!language) return;
+
+        // @ts-ignore
+        i18n.use(initReactI18next).init({
+            ...i18nAppConfig,
+            lng: language,
+        });
+        setLanguageLoaded(true);
+    }, [language]);
+
+    if (!languageLoaded) {
+        return <></>;
+    }
 
     return (
         <TamaguiProvider config={tamaguiConfig} defaultTheme={colorScheme!}>
