+++ src/pages/index.tsx
@@ -10,7 +10,8 @@ import { useAppTranslation } from "src/hooks/useAppTranslation";
 import { useCreatePlanCategory } from "src/hooks/useCreatePlanCategory";
 import { useGooglePlaceSearch } from "src/hooks/useGooglePlaceSearch";
 import { usePwaInstall } from "src/hooks/usePwaInstall";
-import { TranslationNameSpaces, i18nAppConfig } from "src/locales/i18n";
+import { i18nAppConfig } from "src/locales/i18n";
+import { TranslationNameSpaces } from "src/locales/resources";
 import { CreatePlanCategoryList } from "src/view/category/CreatePlanCategoryList";
 import { CreatePlanRangeDialog } from "src/view/category/CreatePlanRangeDialog";
 import { Layout } from "src/view/common/Layout";
